- Updated the emulator
git-svn-id: svn://localhost/mosa@1436 45b04e34-05f1-4d87-a130-40a8491d5c8b

+++ Mosa/EmulatedKernel/MemoryHandler.cs
@@ -78,7 +78,7 @@ namespace Mosa.EmulatedKernel
 		/// <returns>
 		/// 	<c>true</c> if [contains] [the specified address]; otherwise, <c>false</c>.
 		/// </returns>
-		public bool Contains(uint address)
+		public bool Contains(ulong address)
 		{
 			return ((address >= this.Address) && (address < (this.Address + this.Size)));
 		}
